+++ src/compose.ts
@@ -575,13 +575,13 @@ export const composeHandler = ({
 			: []
 
 	const hasBody =
-		hasUnknownContext ||
-		(method !== 'GET' &&
-			method !== 'HEAD' &&
-			hooks.type !== 'none' &&
-			(!!validator.body ||
-				!!hooks.type ||
-				lifeCycleLiteral.some((fn) => isFnUse('body', fn))))
+		method !== 'GET' &&
+		method !== 'HEAD' &&
+		(hasUnknownContext ||
+			(hooks.type !== 'none' &&
+				(!!validator.body ||
+					!!hooks.type ||
+					lifeCycleLiteral.some((fn) => isFnUse('body', fn)))))
 
 	const hasHeaders =
 		hasUnknownContext ||
@@ -781,7 +781,7 @@ export const composeHandler = ({
 		}
 	})
 
-	fnLiteral += hasErrorHandler ? 'try {\n' : ''
+	fnLiteral += hasErrorHandler ? '\n try {\n' : ''
 
 	if (hasTraceSet) {
 		fnLiteral += `\nconst traceDone = Promise.all([`
@@ -1390,25 +1390,25 @@ export const composeHandler = ({
 
 		fnLiteral += `const set = c.set
 
-		if (!set.status || set.status < 300) set.status = 500
+		if (!set.status || set.status < 300) set.status = error?.status || 500
 	`
 
 		const endError = report('error', {
 			unit: hooks.error.length
 		})
 		if (hooks.error.length) {
+			fnLiteral += `
+				c.error = error
+				c.code = error.code ?? error[ERROR_CODE] ?? "UNKNOWN"
+			`
+
 			for (let i = 0; i < hooks.error.length; i++) {
 				const name = `er${i}`
 				const endUnit = report('error.unit', {
 					name: hooks.error[i].name
 				})
 
-				fnLiteral += `\nlet ${name} = handleErrors[${i}](
-					Object.assign(c, {
-						error: error,
-						code: error.code ?? error[ERROR_CODE] ?? "UNKNOWN"
-					})
-				)\n`
+				fnLiteral += `\nlet ${name} = handleErrors[${i}](c)\n`
 
 				if (isAsync(hooks.error[i]))
 					fnLiteral += `if (${name} instanceof Promise) ${name} = await ${name}\n`
@@ -1648,7 +1648,7 @@ export const composeGeneralHandler = (
 			unit: app.event.request.length
 		})
 
-		fnLiteral += `try {\n`
+		fnLiteral += `\n try {\n`
 
 		for (let i = 0; i < app.event.request.length; i++) {
 			const fn = app.event.request[i]
@@ -1806,23 +1806,24 @@ export const composeErrorHandler = (
 		app.event.error.find(isAsync) ? 'async' : ''
 	} function(context, error) {
 		const { set } = context
+
+		context.code = error.code
+		context.error = error
 		`
 	for (let i = 0; i < app.event.error.length; i++) {
 		const handler = app.event.error[i]
 
-		const response = `${isAsync(handler) ? 'await ' : ''}onError[${i}](
-			Object.assign(context, {
-				code: error.code ?? error[ERROR_CODE] ?? 'UNKNOWN',
-				error
-			})
-		)`
+		const response = `${isAsync(handler) ? 'await ' : ''}onError[${i}](context)`
 
 		if (hasReturn(handler.toString()))
-			fnLiteral += `const r${i} = ${response}; if(r${i} !== undefined) return mapResponse(r${i}, set)\n`
+			fnLiteral += `const r${i} = ${response}; if(r${i} !== undefined) {
+				if(set.status === 200) set.status = error.status
+				return mapResponse(r${i}, set)
+			}\n`
 		else fnLiteral += response + '\n'
 	}
 
-	fnLiteral += `if(error.constructor.name === "ValidationError") {
+	fnLiteral += `if(error.constructor.name === "ValidationError" || error.constructor.name === "TransformDecodeError") {
 		set.status = error.status ?? 400
 		return new Response(
 			error.message,