+++ src/error.ts
@@ -1,5 +1,6 @@
 import { Value } from '@sinclair/typebox/value'
 import type { TypeCheck } from '@sinclair/typebox/compiler'
+import { TSchema } from '@sinclair/typebox'
 
 import { StatusMap } from './utils'
 
@@ -87,10 +88,15 @@ export class ValidationError extends Error {
 
 	constructor(
 		public type: string,
-		public validator: TypeCheck<any>,
+		public validator: TSchema | TypeCheck<any>,
 		public value: unknown
 	) {
-		const error = isProduction ? undefined : validator.Errors(value).First()
+		const error = isProduction
+			? undefined
+			: 'Errors' in validator
+			? validator.Errors(value).First()
+			: Value.Errors(validator, value).First()
+
 		const customError = error?.schema.error
 			? typeof error.schema.error === 'function'
 				? error.schema.error(type, validator, value)
@@ -137,9 +143,19 @@ export class ValidationError extends Error {
 		return [...this.validator.Errors(this.value)]
 	}
 
-	get model() {
+	static simplifyModel(validator: TSchema | TypeCheck<any>) {
 		// @ts-ignore
-		return Value.Create(this.validator.schema)
+		const model = 'schema' in validator ? validator.schema : validator
+
+		try {
+			return Value.Create(model)
+		} catch {
+			return model
+		}
+	}
+
+	get model() {
+		return ValidationError.simplifyModel(this.validator)
 	}
 
 	toResponse(headers?: Record<string, any>) {