+++ src/index.ts
@@ -27,7 +27,8 @@ import {
 	mergeLifeCycle,
 	filterGlobalHook,
 	asGlobal,
-	traceBackExtension
+	traceBackExtension,
+	getHostname
 } from './utils'
 
 import {
@@ -85,7 +86,7 @@ import type {
 	ExtensionManager,
 	MapResponse
 } from './types'
-import { t } from './custom-types'
+import { t } from './type-system'
 
 /**
  * ### Elysia Server
@@ -182,7 +183,7 @@ export default class Elysia<
 
 	constructor(config?: Partial<ElysiaConfig<BasePath, Scoped>>) {
 		this.config = {
-			forceErrorEncapsulation: false,
+			forceErrorEncapsulation: true,
 			prefix: '',
 			aot: true,
 			strictPath: false,
@@ -1244,30 +1245,14 @@ export default class Elysia<
 				{}
 			>
 		) => NewElysia
-	): NewElysia extends Elysia<
-		any,
-		infer PluginDecorators,
-		infer PluginDefinitions,
-		infer PluginSchema,
-		any,
-		any
-	>
-		? Elysia<
+	): Elysia<
 				BasePath,
-				Prettify<Decorators & PluginDecorators>,
-				{
-					type: Prettify<
-						Definitions['type'] & PluginDefinitions['type']
-					>
-					error: Prettify<
-						Definitions['error'] & PluginDefinitions['error']
-					>
-				},
-				Prettify<ParentSchema & PluginSchema>,
+				Decorators,
+				Definitions,
+			ParentSchema,
 				Extension,
 				Prettify<Routes & NewElysia['schema']>
 		  >
-		: this
 
 	/**
 	 * ### group
@@ -1299,11 +1284,13 @@ export default class Elysia<
 			prefix: ''
 		})
 		instance.store = this.store
+		instance.getServer = () => this.server
 
 		const isSchema = typeof schemaOrRun === 'object'
 
 		const sandbox = (isSchema ? run! : schemaOrRun)(instance)
 		this.decorators = mergeDeep(this.decorators, instance.decorators)
+
 
 		if (sandbox.event.request.length)
 			this.event.request = [
@@ -1418,23 +1405,14 @@ export default class Elysia<
 				Scoped
 			>
 		) => NewElysia
-	): NewElysia extends Elysia<
-		any,
-		infer PluginDecorators,
-		infer PluginDefinitions,
-		infer PluginSchema,
-		any,
-		any
-	>
-		? Elysia<
-				BasePath,
-				PluginDecorators,
-				PluginDefinitions,
-				PluginSchema,
-				Extension,
-				Prettify<Routes & NewElysia['schema']>
-		  >
-		: this
+	): Elysia<
+	BasePath,
+	Decorators,
+	Definitions,
+	ParentSchema,
+	Extension,
+	Prettify<Routes & NewElysia['schema']>
+>
 
 	/**
 	 * ### guard
@@ -1571,7 +1549,17 @@ export default class Elysia<
 		infer IsScoped
 	>
 		? IsScoped extends true
-			? this
+			? Elysia<
+					BasePath,
+					Decorators,
+					Definitions,
+					ParentSchema,
+					Extension,
+					BasePath extends ``
+						? Routes & NewElysia['schema']
+						: Routes & AddPrefix<BasePath, NewElysia['schema']>,
+					Scoped
+			  >
 			: Elysia<
 					BasePath,
 					{
@@ -1705,10 +1693,11 @@ export default class Elysia<
 			this.lazyLoadModules.push(
 				plugin
 					.then((plugin) => {
-						if (typeof plugin === 'function')
+						if (typeof plugin === 'function') {
 							return plugin(
 								this as unknown as any
 							) as unknown as Elysia
+						}
 
 						if (typeof plugin.default === 'function')
 							return plugin.default(
@@ -1740,6 +1729,39 @@ export default class Elysia<
 				this.lazyLoadModules.push(
 					instance
 						.then((plugin) => {
+							if (plugin instanceof Elysia) {
+								this.compile()
+
+								// Recompile async plugin routes
+								for (const {
+									method,
+									path,
+									handler,
+									hooks
+								} of Object.values(plugin.routes)) {
+									this.add(
+										method,
+										path,
+										handler,
+										mergeHook(
+											hooks as LocalHook<
+												any,
+												any,
+												any,
+												any,
+												any,
+												any
+											>,
+											{
+												error: plugin.event.error
+											}
+										)
+									)
+								}
+
+								return plugin
+							}
+
 							if (typeof plugin === 'function')
 								return plugin(
 									this as unknown as any
@@ -1909,7 +1931,9 @@ export default class Elysia<
 			const run = typeof path === 'function' ? path : handle!
 
 			const handler: Handler<any, any> = async ({ request, path }) =>
-				run(new Request('http://a.cc' + path || '/', request))
+				run(
+					new Request(getHostname(request.url) + path || '/', request)
+				)
 
 			this.all(
 				'/',
@@ -1932,7 +1956,10 @@ export default class Elysia<
 		const length = path.length
 		const handler: Handler<any, any> = async ({ request, path }) =>
 			handle!(
-				new Request('http://a.cc' + path.slice(length) || '/', request)
+				new Request(
+					getHostname(request.url) + path.slice(length) || '/',
+					request
+				)
 			)
 
 		this.all(
@@ -3906,7 +3933,7 @@ export default class Elysia<
 export { Elysia }
 
 export { mapResponse, mapCompactResponse, mapEarlyResponse } from './handler'
-export { t } from './custom-types'
+export { t } from './type-system'
 export { Cookie, type CookieOptions } from './cookie'
 
 export {