+++ test/units/map-early-response.test.ts
@@ -54,7 +54,7 @@ describe('Map Early Response', () => {
 		const response = mapEarlyResponse(body, defaultContext)
 
 		expect(response).toBeInstanceOf(Response)
-		expect(await response?.json()).toEqual(body)
+		expect(await response?.json<any>()).toEqual(body)
 		expect(response?.status).toBe(200)
 	})
 
@@ -87,7 +87,7 @@ describe('Map Early Response', () => {
 		)
 
 		expect(response).toBeInstanceOf(Response)
-		expect(await response?.json()).toEqual(body)
+		expect(await response?.json<any>()).toEqual(body)
 		expect(response?.status).toBe(200)
 	})
 
@@ -106,7 +106,7 @@ describe('Map Early Response', () => {
 		const response = mapEarlyResponse(new Student('Himari'), defaultContext)
 
 		expect(response).toBeInstanceOf(Response)
-		expect(await response?.json()).toEqual({
+		expect(await response?.json<any>()).toEqual({
 			name: 'Himari'
 		})
 		expect(response?.status).toBe(200)
@@ -143,7 +143,7 @@ describe('Map Early Response', () => {
 		)
 
 		expect(response).toBeInstanceOf(Response)
-		expect(await response?.json()).toEqual(body)
+		expect(await response?.json<any>()).toEqual(body)
 		expect(response?.headers.toJSON()).toEqual({
 			...context.headers,
 			'content-type': 'application/json;charset=utf-8'
@@ -155,7 +155,7 @@ describe('Map Early Response', () => {
 		const response = mapEarlyResponse(new Error('Hello'), context)
 
 		expect(response).toBeInstanceOf(Response)
-		expect(await response?.json()).toEqual({
+		expect(await response?.json<any>()).toEqual({
 			name: 'Error',
 			message: 'Hello'
 		})
@@ -169,7 +169,7 @@ describe('Map Early Response', () => {
 
 		expect(response).toBeInstanceOf(Response)
 		expect(await response?.text()).toEqual('Shiroko')
-		expect(response?.headers.toJSON()).toEqual(headers)
+		expect(response?.headers.toJSON()).toEqual(headers as any)
 	})
 
 	it('map Response and merge Headers', async () => {
@@ -249,10 +249,8 @@ describe('Map Early Response', () => {
 		})
 		expect(response).toBeInstanceOf(Response)
 		expect(await response?.text()).toEqual('Hina')
-		expect(response?.headers.toJSON()).toEqual({
-			name: 'Sorasaki Hina',
-			'set-cookie': ['name=hina']
-		})
+		expect(response?.headers.get('name')).toEqual('Sorasaki Hina')
+		expect(response?.headers.getAll('set-cookie')).toEqual(['name=hina'])
 	})
 
 	it('set multiple cookie', async () => {
@@ -272,10 +270,11 @@ describe('Map Early Response', () => {
 		})
 		expect(response).toBeInstanceOf(Response)
 		expect(await response?.text()).toEqual('Hina')
-		expect(response?.headers.toJSON()).toEqual({
-			name: 'Sorasaki Hina',
-			'set-cookie': ['name=hina', 'affiliation=gehenna']
-		})
+		expect(response?.headers.get('name')).toEqual('Sorasaki Hina')
+		expect(response?.headers.getAll('set-cookie')).toEqual([
+			'name=hina',
+			'affiliation=gehenna'
+		])
 	})
 
 	it('map Passthrough', async () => {